Unify GitReference across stores
