+++ crates/store/src/source.rs
@@ -22,8 +22,6 @@ use crate::stores::providers::git::{GitAuth, GitReference};
 
 #[cfg(feature = "github")]
 use crate::stores::github::GitHubStore;
-#[cfg(feature = "github")]
-use crate::stores::github_file_operations::GitReference as GitHubGitReference;
 
 /// Type of extension store with associated data
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -383,7 +381,7 @@ impl ExtensionSource {
             } => {
                 let github_store = GitHubStore::builder(owner.clone(), repo.clone())
                     .auth(auth.clone())
-                    .reference(GitHubGitReference::from_git_reference(&reference))
+                    .reference(reference.clone())
                     .cache_dir(cache_dir.clone())
                     .name(self.name.clone())
                     .build()
@@ -439,7 +437,7 @@ impl ExtensionSource {
             } => {
                 let github_store = GitHubStore::builder(owner.clone(), repo.clone())
                     .auth(auth.clone())
-                    .reference(GitHubGitReference::from_git_reference(&reference))
+                    .reference(reference.clone())
                     .cache_dir(cache_dir.clone())
                     .name(self.name.clone())
                     .writable()
@@ -495,7 +493,7 @@ impl ExtensionSource {
             } => {
                 let github_store = GitHubStore::builder(owner.clone(), repo.clone())
                     .auth(auth.clone())
-                    .reference(GitHubGitReference::from_git_reference(&reference))
+                    .reference(reference.clone())
                     .cache_dir(cache_dir.clone())
                     .name(self.name.clone())
                     .build()
@@ -555,7 +553,7 @@ pub async fn create_readable_store_from_source(
         } => {
             let github_store = GitHubStore::builder(owner.clone(), repo.clone())
                 .auth(auth.clone())
-                .reference(GitHubGitReference::from_git_reference(&reference))
+                .reference(reference.clone())
                 .cache_dir(cache_dir.clone())
                 .name(source.name.clone())
                 .build()
