+++ crates/store/src/stores/github.rs
@@ -21,7 +21,8 @@ use crate::publish::{
 };
 use crate::store_manifest::StoreManifest;
 use crate::stores::file_operations::FileBasedProcessor;
-use crate::stores::github_file_operations::{GitHubFileOperations, GitReference};
+use crate::stores::github_file_operations::GitHubFileOperations;
+use crate::stores::providers::git::GitReference;
 use crate::stores::providers::git::{GitAuth, GitWriteConfig};
 use crate::stores::traits::{BaseStore, CacheableStore, ReadableStore, WritableStore};
